* Added preliminary code for rendering regions * Added various utility methods to classes * improved the locking detection for the threadingmanager

+++ src/gpvm/map/TileData.java
@@ -67,7 +67,7 @@ public final class TileData {
    *
    * @return A list of all field names.
    */
-  public int[] getValueNames() {
+  public long[] getValueNames() {
     return values.keySet();
   }
